requirements.txt
appdirs==1.4.4
asgiref==3.9.1
attrs==25.3.0
cattrs==25.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.5
exceptiongroup==1.3.0
fonttools==4.59.1
geojson-rewind==1.1.0
geomet==1.1.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.2.6
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pycountry==24.6.1
pygbif==0.6.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.5
requests-cache==1.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
url-normalize==2.2.1
urllib3==2.5.0
